move grid from battle-action-reducer to snapshot-reducer
